apps/api/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pocketflow
pyyaml